Fix the issue of bug in the Weather Report Api call since the error response doesn't have key "message"
Change the version of <source> and <target> inside pom.xml to make whole project more compatible.

In order to make the chat system more interesting and interactive, a joke button has been added. If the chat loses topic, the chatter can click the "Get joke" button to get a joke with a question and answer. The chatter can ask questions first, allowing other chatters to participate in guessing the answers, making the chat interactive. Moreover, every time the "get joke" button is clicked, the program can give a new joke, which avoids the repetition of jokes and is random, adding a certain sense of surprise to the user and making the chat system more interesting.